requirements.txt
accelerate==0.19.0
black==23.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
diffusers==0.16.1
filelock==3.12.0
fsspec==2023.5.0
huggingface-hub==0.14.1
idna==3.4
importlib-metadata==6.6.0
Jinja2==3.1.2
MarkupSafe==2.1.2
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.1
numpy==1.24.3
opencv-python==4.7.0.72
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.5.1
psutil==5.9.5
PyYAML==6.0
regex==2023.5.5
requests==2.30.0
six==1.16.0
sympy==1.12
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.29.2
typing_extensions==4.5.0
urllib3==2.0.2
zipp==3.15.0
